Move remarshal into _internal
