+++ src/globus_sdk/services/compute/client.py
@@ -4,8 +4,8 @@ import logging
 import typing as t
 
 from globus_sdk import GlobusHTTPResponse, client
+from globus_sdk._internal.remarshal import strseq_listify
 from globus_sdk._missing import MISSING, MissingType
-from globus_sdk._remarshal import strseq_listify
 from globus_sdk._types import UUIDLike
 from globus_sdk.scopes import ComputeScopes
