+++ src/globus_sdk/services/gcs/client.py
@@ -5,8 +5,8 @@ import uuid
 
 from globus_sdk import client, exc, paging, response
 from globus_sdk._internal.classprop import classproperty
+from globus_sdk._internal.remarshal import commajoin
 from globus_sdk._missing import MISSING, MissingType
-from globus_sdk._remarshal import commajoin
 from globus_sdk._types import UUIDLike
 from globus_sdk._utils import slash_join
 from globus_sdk.authorizers import GlobusAuthorizer
